refactor: extract vdisk query params hook
