+++ src/containers/VDiskPage/VDiskPage.tsx
@@ -4,9 +4,6 @@ import {Flex, Icon, Label, Tab, TabList, TabProvider} from '@gravity-ui/uikit';
 import {skipToken} from '@reduxjs/toolkit/query';
 import {isNil} from 'lodash';
 import {Helmet} from 'react-helmet-async';
-import {Redirect, useLocation} from 'react-router-dom';
-import {StringParam, useQueryParams} from 'use-query-params';
-import {z} from 'zod';
 
 import {EntityPageTitle} from '../../components/EntityPageTitle/EntityPageTitle';
 import {ResponseError} from '../../components/Errors/ResponseError';
@@ -28,66 +25,25 @@ import {parseVdiskId} from '../../utils/dataFormatters/dataFormatters';
 import {VDISK_LABEL_CONFIG} from '../../utils/disks/constants';
 import {getDataSeverityColor} from '../../utils/disks/helpers';
 import {useAutoRefreshInterval, useTypedDispatch} from '../../utils/hooks';
-import {useIsViewerUser} from '../../utils/hooks/useIsUserAllowedToMakeChanges';
 import {useAppTitle} from '../App/AppTitleContext';
 import {PaginatedStorage} from '../Storage/PaginatedStorage';
 
 import {VDiskStorageDetails} from './VDiskStorageDetails';
 import {VDiskTablets} from './VDiskTablets';
 import {vDiskPageKeyset} from './i18n';
+import {useVDiskQueryParams} from './useVDiskQueryParams';
 
 import './VDiskPage.scss';
 
 const vDiskPageCn = cn('ydb-vdisk-page');
 
-const VDISK_TABS_IDS = {
-    storage: 'storage',
-    tablets: 'tablets',
-} as const;
-
-const VDISK_PAGE_TABS = [
-    {
-        id: VDISK_TABS_IDS.storage,
-        get title() {
-            return vDiskPageKeyset('storage');
-        },
-    },
-    {
-        id: VDISK_TABS_IDS.tablets,
-        get title() {
-            return vDiskPageKeyset('tablets');
-        },
-    },
-];
-
-const vDiskTabSchema = z.nativeEnum(VDISK_TABS_IDS).catch(VDISK_TABS_IDS.storage);
-
 export function VDiskPage() {
     const dispatch = useTypedDispatch();
     const getVDiskPagePath = useVDiskPagePath();
-    const location = useLocation();
 
     const containerRef = React.useRef<HTMLDivElement>(null);
 
-    const [{nodeId, vDiskId: vDiskIdParam, activeTab, database: databaseParam}] = useQueryParams({
-        nodeId: StringParam,
-        vDiskId: StringParam,
-        activeTab: StringParam,
-        database: StringParam,
-    });
-    const database = databaseParam ?? undefined;
-
-    const requestedVDiskTab = vDiskTabSchema.parse(activeTab);
-    const isViewerUser = useIsViewerUser();
-    const {vDiskTab, vDiskTabs} = React.useMemo(() => {
-        const availableTabs = isViewerUser
-            ? VDISK_PAGE_TABS
-            : VDISK_PAGE_TABS.filter(({id}) => id !== VDISK_TABS_IDS.tablets);
-        const availableActiveTab =
-            availableTabs.find(({id}) => id === requestedVDiskTab)?.id ?? VDISK_TABS_IDS.storage;
-
-        return {vDiskTab: availableActiveTab, vDiskTabs: availableTabs};
-    }, [isViewerUser, requestedVDiskTab]);
+    const {nodeId, vDiskId: vDiskIdParam, database, vDiskTab, vDiskTabs} = useVDiskQueryParams();
     const newStorageViewEnabled = useNewStorageViewEnabled();
 
     const [autoRefreshInterval] = useAutoRefreshInterval();
@@ -136,17 +92,6 @@ export function VDiskPage() {
     const {GroupID} = resolvedVDiskId || {};
 
     const vDiskId = vDiskData?.StringifiedId || (loading ? undefined : vDiskIdParam);
-    const redirectLocation = React.useMemo(() => {
-        if (vDiskTab === requestedVDiskTab) {
-            return undefined;
-        }
-
-        const searchParams = new URLSearchParams(location.search);
-        searchParams.set('activeTab', vDiskTab);
-
-        return {...location, search: searchParams.toString()};
-    }, [location, requestedVDiskTab, vDiskTab]);
-
     const {appTitle} = useAppTitle();
 
     const renderHelmet = () => {
@@ -338,10 +283,6 @@ export function VDiskPage() {
             </React.Fragment>
         );
     };
-
-    if (redirectLocation) {
-        return <Redirect to={redirectLocation} />;
-    }
 
     return (
         <div className={vDiskPageCn(null)} ref={containerRef}>
